Added ReleaseTest configuration to allow running tests against optimized build.
It is not possible to run the test suite with the ENABLE_TESTABILITY flag set to NO. It is also not recommended by Apple to set the ENABLE_TESTABILITY flag to YES for release builds. In order to still be able to run the optimized build of the framework against the test suite, the ReleaseTest configuration was created. It optimizes the framework as well as sets the ENABLE_TESTABILITY flag to YES. This allows the test suite to run on Travis-CI in the ReleaseTest configuration, and does not affect Carthage installation.