[Issue 1785] Fixed issue where adapt error was incorrectly sent to retrier.
The AdaptError is only for internal use and should never be sent back to any public APIs such as the retrier. Previously, the AdaptError was being incorrectly sent to the retrier instead of the underlying error. This commit fixes the issue and also adds tests to ensure this is working properly.