[PR #2047] Added missing charset to the JSONEncoding encode API.
The charset is needed to make sure, that the server uses the correct charset for decoding. This in necessary since JSON can be encoded in various forms of unicode.

RFC 7158 - Section 8.1:

JSON text SHALL be encoded in UTF-8, UTF-16, or UTF-32.  The default encoding is UTF-8, and JSON texts that are encoded in UTF-8 are interoperable in the sense that they will be read successfully by the maximum number of implementations; there are many implementations that cannot successfully read texts in other encodings (such as UTF-16 and UTF-32).