Fix Reachability Safety (#3684)
### Issue Link :link:
#3676

### Goals :soccer:
This PR updates our integration of `SCNetworkReachabilityContext` to use
a weak box class and fully integrate with the framework's retain and
release closures. This follows the same logic as
[Reachability.swift](https://github.com/ashleymills/Reachability.swift/blob/a81b7367f2c46875f29577e03a60c39cdfad0c8d/Sources/Reachability.swift#L190).

### Implementation Details :construction:
This PR adds a `WeakManager` type to wrap a weak reference to the
`NetworkReachabilityManager` and updates the context memory management
closures to properly retain and release the values.

### Testing Details :mag:
No tests added, wasn't able to find a way to trigger the reported crash.
Since this logic matches that of the other popular reachability library
I'm confident it's correct.