Allow Automatic Cancellation By Default (#3757)
### Issue Link :link:
#3727

### Goals :soccer:
This PR changes Alamofire's default behavior to better follow user
expectations around automatically cancellation when the enclosing async
context is cancelled.

This PR also updates some other async behaviors for more reliable tests
and behaviors.

### Implementation Details :construction:
Previously, Alamofire defaulted `shouldAutomaticallyCancel` to `true` in
order to follow what was then anticipated user expectations as well as
correct behavior for async APIs. However, many of the Swift Concurrency
cancellation behaviors that prompted this default turned out to be bugs.
Now that the correct behavior has been around for a while it's important
for Alamofire to fix its own behavior to match.

### Testing Details :mag:
Updated tests for new default and added tests when it's turned off.