Remove @propertyWrapper From Protected (#3777)
### Goals :soccer:
Use of a property wrapper is not thread-safe beyond access of the
`wrappedValue` so for a while now Alamofire has had the implicit rule of
always accessing `Protected` values through the wrapper using `$`. We
can make this completely safe by removing the `@propertyWrapper`
conformance altogether and always forcing direct use of `Protected`.
This is somewhat less ergonomic but should be safer in the long run.