Add encodeIfPresent Support to URLEncodedFormEncoder (#3779)
### Issue Link :link:
Fixes #3778

### Goals :soccer:
The encoding of synthesized `Encodable` values is different than
`[String: String?]` so we were missing coverage for `encodeIfPresent`,
causing `Optional` values to be dropped. This PR adds the missing
support.

### Implementation Details :construction:
Added all of the `encodeIfPresent` overloads to
`_URLEncodedFormEncoder.KeyedContainer` so that the protocol uses the
custom implementation.

### Testing Details :mag:
Tests added for `Encodable` values with optional properties.