Add WebSocketRequest Under @_spi(WebSocket) (#3455)
### Goals :soccer:
This PR adds `WebSocketRequest`, an Alamofire wrapper around
`URLSessionWebSocketTask`, as well as assorted functionality for dealing
with WebSockets. This includes:

- Automatic listening: no more manual `listen` calls on the task.
- Automatic ping support: ping the server automatically given an
interval.
- Response serialization support: parse messages into objects, including
using `Decodable`.
- Concurrency support: full streams of events and messages.

### Implementation Details :construction:
`WebSocketRequest` follows most of Alamofire's existing patterns. There
are a few differences:

- Closure vs. cancellation: web sockets can close expectedly, yet this
closure produces an error from `URLSession`. This had to be customized
in order to ignore the error and allow the user to close a connection
with no additional issue.
- Like `DataStreamRequest`, `WebSocketRequest` deals with incoming
streams of data. However, that data is already framed and then partially
decoded into `Message`s by `URLSession`, so serialization had to be
built at that level.
- Unlike `DataStreamRequest`, `WebSocketRequest` can send `Message` back
up the connection. This requires some queuing on the request side to
ensure messages sent before the connection is established are properly
stored for sending.

### Testing Details :mag:
Full suite of `WebSocketTests` added, as well as the concurrency
versions. Firewalk was updated with several new `websocket` cases to
support scenarios like an echo server or expected ping count.

P.S. This feature was delayed so long due to the many hours of
frustration I had to deal with when Apple shipped a message dropping bug
in iOS 14.5 - 14.7. In trying to test the socket connections I would
connect to the test server, have it reply with a message, and then close
the connection. Shockingly that meant that *sometimes* the messages
wouldn't come through. But sometimes they would. Investigating other web
socket clients and inspecting network traffic with Wireshark revealed
that the messages were being sent by the server but the client just
wasn't receiving them. Further investigation showed that delaying the
close of the connection by 60ms or so let the client receive the message
reliably. After many hours I finally had reliable tests! Only to then
test on the iOS 15 betas and see that I didn't need the delay at all,
the tests were reliable without it! Turns out it was just a bug in iOS
14.5 - 14.7. I was so disgusted by this fact that I gave up on the
feature for years. I'll call out this bug in the documentation.